rice-classifier-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.19.0
numpy==1.26.4
pillow==9.5.0 